requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
psycopg2-binary
pyyaml
pytest
lxml